Fixed app structure
