+++ lib/core/providers/auth.dart
