+++ lib/features/admin/widgets/admin_app_bar.dart
@@ -2,8 +2,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gptmoe/main.dart';
-import 'package:gptmoe/state/theme_state_notifier.dart';
-import 'package:gptmoe/common.dart';
+import 'package:gptmoe/core/utils/state/theme_state_notifier.dart';
+import 'package:gptmoe/core/utils/common.dart';
 
 class AdminAppBar {
   static final List<String> _tabs = [
@@ -77,9 +77,7 @@ class ThemeIconButton extends ConsumerWidget {
         onPressed: () {
           ref.read(themeStateNotifierProvider.notifier).changeTheme();
         },
-        icon: Icon(isDarkState == true
-            ? Icons.nightlight
-            : Icons.nightlight_outlined));
+        icon: Icon(isDarkState == true ? Icons.nightlight : Icons.nightlight_outlined));
   }
 }
