+++ lib/features/login/pages/login_page.dart
@@ -1,0 +1,25 @@
+import 'package:flutter_firebase_auth/login_screen.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter/material.dart';
+
+class LoginPage extends ConsumerWidget {
+  const LoginPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return const Scaffold(
+      body: LoginScreen(
+        "GPT Moe",
+        "Log in",
+        {
+          "loginGitHub": true,
+          "loginGoogle": true,
+          "loginEmail": true,
+          "loginSSO": true,
+          "loginAnonymous": true,
+          "signupOption": true,
+        },
+      ),
+    );
+  }
+}
