+++ lib/resources/widgets/app_bar.dart
@@ -2,8 +2,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gptmoe/main.dart';
-import 'package:gptmoe/state/theme_state_notifier.dart';
-import 'package:gptmoe/common.dart';
+import 'package:gptmoe/core/utils/state/theme_state_notifier.dart';
+import 'package:gptmoe/core/utils/common.dart';
 
 class MyAppBar {
   static final List<String> _tabs = [
@@ -12,10 +12,8 @@ class MyAppBar {
 
   static PreferredSizeWidget getBar(BuildContext context, WidgetRef ref) {
     return AppBar(
-      automaticallyImplyLeading:
-          (MediaQuery.of(context).size.width < kWideScreenWidth) ? true : false,
-      leadingWidth:
-          (MediaQuery.of(context).size.width < kWideScreenWidth) ? null : 100,
+      automaticallyImplyLeading: (MediaQuery.of(context).size.width < kWideScreenWidth) ? true : false,
+      leadingWidth: (MediaQuery.of(context).size.width < kWideScreenWidth) ? null : 100,
       leading: (MediaQuery.of(context).size.width < kWideScreenWidth)
           ? null
           : const Padding(
@@ -75,8 +73,6 @@ class ThemeIconButton extends ConsumerWidget {
         onPressed: () {
           ref.read(themeStateNotifierProvider.notifier).changeTheme();
         },
-        icon: Icon(isDarkState == true
-            ? Icons.nightlight
-            : Icons.nightlight_outlined));
+        icon: Icon(isDarkState == true ? Icons.nightlight : Icons.nightlight_outlined));
   }
 }
