+++ lib/resources/widgets/controls/doc_field_text_edit.dart
